+++ bot/migrations/2019-04-03-060909_aliases/up.sql
@@ -1,0 +1,6 @@
+CREATE TABLE aliases (
+    channel VARCHAR NOT NULL,
+    name VARCHAR NOT NULL,
+    text TEXT NOT NULL,
+    PRIMARY KEY (channel, name)
+);