+++ bot/src/lib.rs
@@ -7,7 +7,6 @@ extern crate warp;
 
 pub const VERSION: &'static str = env!("CARGO_PKG_VERSION");
 
-pub mod aliases;
 pub mod bus;
 mod command;
 pub mod config;